+++ src/types.d.ts
@@ -1,11 +1,11 @@
-interface ComponentHanlder {
+interface ComponentHandler {
   name: string
   instance: any
   options?: any
 }
 
 export interface RemarkMDCOptions {
-  components?: ComponentHanlder[]
+  components?: ComponentHandler[]
   experimental?: {
     autoUnwrap?: boolean
     componentCodeBlockProps?: boolean